# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ottosfa/v2.3/Province/": {
            "get": {
                "description": "Province - list by country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Province V2.3"
                ],
                "summary": "Province - list by country",
                "operationId": "Province - list by country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/action-types": {
            "get": {
                "description": "Admin - ActionType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin V2.3"
                ],
                "summary": "Admin - ActionType",
                "operationId": "Admin - ActionType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/company-codes": {
            "get": {
                "description": "Company Codes - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Codes V2.3"
                ],
                "summary": "Company Codes - List",
                "operationId": "Company Codes - List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/country/": {
            "get": {
                "description": "Country - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country V2.3"
                ],
                "summary": "Country - List",
                "operationId": "Country - List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/gender/": {
            "get": {
                "description": "Gender - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gender V2.3"
                ],
                "summary": "Gender - List",
                "operationId": "Gender - List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/todolist/merchant-detail": {
            "post": {
                "description": "Merchant Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList V2.3"
                ],
                "summary": "Merchant Detail",
                "operationId": "merchant-detail",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MerchantDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/todolist/merchant-list": {
            "post": {
                "description": "Merchant List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList V2.3"
                ],
                "summary": "Merchant List",
                "operationId": "merchant-list",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MerchantListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/upload-image": {
            "post": {
                "description": "Upload Image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "UploadImage",
                "operationId": "Upload Image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miniomodels.UploadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/village/list": {
            "get": {
                "description": "Village - List by district",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Village V2.3"
                ],
                "summary": "Village - List by district",
                "operationId": "Village - List by district",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/city/list": {
            "post": {
                "description": "City - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "City - List",
                "operationId": "City - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City - List Response EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dbmodels.Cities"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/district/list": {
            "post": {
                "description": "District - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "District - List",
                "operationId": "District - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "District - List Response EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dbmodels.District"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/institution/list": {
            "post": {
                "description": "Institution - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Institution - List",
                "operationId": "Institution - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/merchant-business-type/list": {
            "get": {
                "description": "MerchantBusinessType - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "MerchantBusinessType - List",
                "operationId": "MerchantBusinessType - List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/merchant-category/list": {
            "get": {
                "description": "MerchantCategory - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "MerchantCategory - List",
                "operationId": "MerchantCategory - List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/merchant-group/list": {
            "post": {
                "description": "MerchantGroup - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "MerchantGroup - List",
                "operationId": "MerchantGroup - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MerchantGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/merchant-type/list": {
            "get": {
                "description": "MerchantType - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "MerchantType - List",
                "operationId": "MerchantType - List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/sales-retail/list": {
            "get": {
                "description": "SalesRetail - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "SalesRetail - List",
                "operationId": "SalesRetail - List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dbmodels.Cities": {
            "type": "object",
            "properties": {
                "city_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "province_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "miniomodels.UploadReq": {
            "type": "object",
            "properties": {
                "bucketName": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "nameFile": {
                    "type": "string"
                }
            }
        },
...
        "models.MerchantDetailReq": {
            "type": "object",
            "properties": {
                "merchant_id": {
                    "type": "string"
                }
            }
        },
        "models.MerchantGroupReq": {
            "type": "object",
            "properties": {
                "merchant_type": {
                    "type": "string",
                    "example": "UMKM"
                }
            }
        },
        "models.MerchantListReq": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                }
            }
        },
...
        "models.SearchReq": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string",
                    "example": "IDM"
                }
            }
        },
...
    }
}